mcp/pyproject.toml
[project]
name = "caof-mcp"
version = "0.1.0"
description = "CAOF MCP server — task execution via Model Context Protocol"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
caof-mcp = "server:main"
